Motion Sensing with Conditions and Variable Dimming
Controls Lights based on a Motion Sensor.  You can use Switches, that when on, suspend the current state of the lights and they will not change as long as one of the switches are on.  The app also has the ability to variably set dimming lights based on the lux level of a selected lux sensor.  You can also select which modes you wish the app to run in.